Initial refactor - compiles but needs fixing etc.
