+++ OpenSim/Simulation/Model/Blankevoort1991Ligament.h
@@ -24,9 +24,7 @@
  * -------------------------------------------------------------------------- */
 
 #include <OpenSim/Simulation/Model/Force.h>
-//#include <OpenSim/Simulation/Model/GeometryPath.h>
-#include <OpenSim/Simulation/Model/PointBasedPath.h>
-#include <OpenSim/Simulation/Model/FunctionBasedPath.h>
+#include <OpenSim/Simulation/Model/GeometryPath.h>
 
 namespace OpenSim {
